AMDGPU API Reference

Synchronization

AMDGPU.synchronize — Function

synchronize(stream::HIPStream = stream(); blocking::Bool = true)

Blocks until all kernels currently executing on stream have completed.

If there are running HostCalls, then it non-blocking synchronization is required which can be done with blocking=false keyword. Additionally, it stops any running global hostcall afterwards. Note, that non-blocking synchronization is slower than blocking.

AMDGPU.@sync — Macro

@sync ex

Run expression ex on currently active stream and synchronize the GPU on that stream afterwards.

See also: synchronize.

AMDGPU.device_synchronize — Function

Blocks until all kernels on all streams have completed. Uses currently active device.

Device code API

Thread indexing

HSA nomenclature

AMDGPU.Device.workitemIdx — Function

workitemIdx()::ROCDim3

Returns the work item index within the work group. See also: threadIdx

AMDGPU.Device.workgroupIdx — Function

workgroupIdx()::ROCDim3

Returns the work group index. See also: blockIdx

AMDGPU.Device.workgroupDim — Function

workgroupDim()::ROCDim3

Returns the size of each workgroup in workitems. See also: blockDim

AMDGPU.Device.gridItemDim — Function

gridItemDim()::ROCDim3

Returns the size of the grid in workitems. This behaviour is different from CUDA where gridDim gives the size of the grid in blocks.

AMDGPU.Device.gridGroupDim — Function

gridGroupDim()::ROCDim3

Returns the size of the grid in workgroups. This is equivalent to CUDA's gridDim.

CUDA nomenclature

Use these functions for compatibility with CUDA.jl.

AMDGPU.Device.threadIdx — Function

threadIdx()::ROCDim3

Returns the thread index within the block. See also: workitemIdx

AMDGPU.Device.blockIdx — Function

blockIdx()::ROCDim3

Returns the block index within the grid. See also: workgroupIdx

AMDGPU.Device.blockDim — Function

blockDim()::ROCDim3

Returns the dimensions of the block. See also: workgroupDim

Synchronization

AMDGPU.Device.sync_workgroup — Function

sync_workgroup()

Waits until all wavefronts in a workgroup have reached this call. This function is meant to be used inside kernels.