Queues

Similar to CUDA streams, ROCm has the concept of queues, which are buffers used to instruct the GPU hardware which kernels to launch. ROCm queues are synchronous, like CUDA streams.

Each device has a default queue associated, which is accessible with AMDGPU.queue.

To specify which queue to launch a kernel on:

Using AMDGPU.queue!, which will execute given function and reset to the original queue after completion:

q = AMDGPU.ROCQueue()
x = AMDGPU.queue!(() -> AMDGPU.ones(Float32, 16), q)

Using queue argument to @roc macro:

q = AMDGPU.ROCQueue()
@roc queue=q kernel(...)

Queues also have an inherent priority, which allows control of kernel submission latency and on-device scheduling preference with respect to kernels submitted on other queues. There are three priorities: normal (the default), low, and high priority.

Priority of the default queue can be set with AMDGPU.priority!. Alternatively, it can be set at queue creation time:

low_prio_queue = ROCQueue(; priority=:low)
high_prio_queue = ROCQueue(; priority=:high)
normal_prio_queue = ROCQueue(; priority=:normal) # or just omit "priority"

To get kernels which are currently executing on a given queue, use AMDGPU.active_kernels. It will return a Vector{ROCKernelSignal}, which can be inspected to determine how many (and which) kernels are executing.

If a kernel gets "stuck" and locks up the GPU (noticeable with 100% GPU usage in rocm-smi) you can kill it and all other kernels associated with the queue it is running on with AMDGPU.Runtime.kill_queue!(queue). This can be "safely" done to the default queue (obtained via AMDGPU.queue), since default queues are recreated as-needed.

AMDGPU.queue — Function

queue()::ROCQueue

Get task-local default queue for the currently active device.

AMDGPU.queue! — Function

queue!(f::Base.Callable, queue::ROCQueue)

Change default queue, execute given function f and revert back to the original queue.

Returns

Return value of the function f.

AMDGPU.priority! — Function

priority!(priority::Symbol)

Change the priority of the default queue. Accepted values are :normal (the default), :low and :high.

priority!(f::Base.Callable, priority::Symbol)

Chnage the priority of default queue, execute f and revert to the original priority. Accepted values are :normal (the default), :low and :high.

Returns

Return value of the function f.

AMDGPU.active_kernels — Function

active_kernels(queue::ROCQueue = queue()) -> Vector{ROCKernelSignal}

Returns the set of actively-executing kernels on queue.

AMDGPU.Runtime.ROCQueue — Type

ROCQueue(; priority::Symbol=:normal, pooled::Bool=false)

Create an HSA queue on the currently active device.

Note

Users are encouraged to use this method, instead of manually providing device since this one correctly handles device changes.

ROCQueue(device::ROCDevice; priority::Symbol=:normal, pooled::Bool=false)

Create an HSA queue which will be used to instruct GPU hardware which kernels to launch.

Each queue, spawns an error monitoring thread that's responsible for actually waiting on kernels and performing a cleanup after kernel finished its execution.

Oversubscribed Command Queues in GPUs

Be careful, with the number of HSA queues in use. When the number of allocated HSA queues is greater than the number of hardware queues, the GPU wastes significant time rotating between all allocated queues in search of ready tasks.

Arguments:

device::ROCDevice: Device on which to create queue.

priority::Symbol: Queue's priority. Can be :normal, :low, :high.

pooled::Bool: Whether to use pool when creating queues. When true, queues are drawn from it on creation and returned to pool instead of destroyed.

AMDGPU.Runtime.set_queue_pool_size! — Function

set_queue_pool_size!(nums::NTuple{3, Int})

Set HSA queue pool max size for each priority. Restart Julia session for the changes to take effect.

Arguments:

nums::NTuple{3, Int}: Maximum number of queues for :normal, :low and :high priority. Providing 0 for specific priority, disables pool for it.

AMDGPU.Runtime.kill_queue! — Function

kill_queue!(queue::ROCQueue)

Kill queue and propagate queue error to all waiter signals in case if there is one.

If queue is in the pool, it will be removed from it.

Note

No need to manually call this function during regular use, it will be called automatically from ROCQueue finalizer.

Signals

Unlike CUDA, ROCm kernels are tracked by an associated signal, which is created and returned by @roc, and is waited on to track kernel completion. Signals may also be used for manual synchronization (since they work for CPUs and GPUs equally well). CPU usage is done with the HSA.signal_* functions, and GPU usage is done with the device_signal_* and hostcall_device_signal_* functions. For most signalling needs, consider using a hostcall instead.

If custom signal handling is desired, signals can be manually constructed and passed to @roc:

# A kernel which waits on all signals in `sigs`
function multi_wait(sigs)
    for i in 1:length(sigs)
        AMDGPU.Device.hostcall_device_signal_wait(sigs[i], 0)
    end
    nothing
end

# Create a set of signals
sigs = [ROCSignal() for i in 1:10]
# Get the device-safe signal handles
_sigs = ROCArray(map(sig->sig.signal, sigs))

# Launch multi-waiter ahead of time; this will block on the device
final_sig = @roc multi_wait(_sigs)

# Associate kernels with signals
for sig in sigs
    @roc signal=sig identity(nothing)
end

# Wait on the multi-waiter
wait(final_sig)