Devices/Agents

In AMDGPU, all GPU devices (also known as "agents" in HSA parlance) are auto-detected by the runtime, if they're supported. There are three classes of devices: CPU, GPU, and DSP. In AMDGPU, we only support compilation and execution on GPU devices, so we will henceforth limit discussion to those; however, you may see a kind Symbol available in the APIs of many device access functions, which defaults to :gpu, but could also be :cpu or :dsp.

AMDGPU maintains a global default device. The default device is relevant for all kernel and GPUArray operations; if one is not specified via @roc or an equivalent interface, then the default device is used for those operations, which affects compilation and kernel launch.

The default device is accessible via AMDGPU.get_default_agent(). This function returns an HSAAgent, which is a handle that references the device. The list of available devices can be queried with AMDGPU.get_agents() to get a list of all known and potentially usable devices. If you have an HSAAgent object, you can also switch the default device via AMDGPU.set_default_agent!(agent).

Additionally, devices have an associated numeric ID. The default device ID can be queried with AMDGPU.device(), which returns an Int. This value is bounded between 1 and length(AMDGPU.get_agents()), and device 1 is the default device when AMDGPU is first loaded. The default device may be switched with AMDGPU.device!(idx).

AMDGPU.get_agents — Function

get_agents() -> Vector{HSAAgent}

Returns the list of HSA agents available on the system.

AMDGPU.get_default_agent — Function

get_default_agent() -> HSAAgent

Returns the default agent, which is used for all kernel and array operations when one is not explicitly specified. May be changed with set_default_agent!.

AMDGPU.set_default_agent! — Function

set_default_agent!(agent::HSAAgent) -> HSAAgent

Sets the default agent to agent. See get_default_agent for more details.

AMDGPU.device — Function

device(kind::Symbol=:gpu) -> Int

Returns the numeric ID of the current default device, which is in the range of 1:length(AMDGPU.get_agents(kind)). This number should be stable for all processes on the same node, so long as any device filtering is consistently applied (such as ROCR_VISIBLE_DEVICES). The device! function accepts the same numeric ID that is produced by this function.

AMDGPU.device! — Function

device!(idx::Integer, kind::Symbol=:gpu)

Sets the default device to AMDGPU.get_agents(kind)[idx]. See device for details on the numbering semantics.