Devices/Agents

In AMDGPU, all GPU devices (also known as "agents" in HSA parlance) are auto-detected by the runtime, if they're supported. There are three classes of devices:

CPU

GPU

DSP

In AMDGPU, we only support compilation and execution on GPU devices, so we will henceforth limit discussion to those; however, you may see a kind Symbol available in the APIs of many device access functions, which defaults to :gpu, but could also be :cpu or :dsp.

AMDGPU maintains a global default device. The default device is relevant for all kernel and GPUArray operations; if one is not specified via @roc or an equivalent interface, then the default device is used for those operations, which affects compilation and kernel launch.

Task-Local Storage

Since AMDGPU.jl relies on Task-Local Storage, this means that default devices are default only within a given task. Other tasks migh have different default devices if the user switched them.

The default device is accessible via AMDGPU.device(). This function returns a ROCDevice, which is a handle that references the device. The list of available devices can be queried with AMDGPU.devices to get a list of all known and potentially usable devices.

If you have a ROCDevice object, you can also switch the default device via AMDGPU.device!. This will switch it only within the task it is called from.

To select default device for newly created tasks, use AMDGPU.default_device!.

Additionally, devices have an associated numeric ID. The default device ID can be queried with AMDGPU.default_device_id, which returns an Int. This value is bounded between 1 and length(AMDGPU.devices()), and device 1 is the default device when AMDGPU is first loaded.

AMDGPU.devices — Function

devices(kind::Symbol = :gpu)

Get list of all devices of the given kind. kind can be :cpu, :gpu or :dsp, although AMDGPU.jl supports execution only on :gpu devices.

AMDGPU.device — Function

device()::ROCDevice

Get currently active device. This device is used when launching kernels via @roc.

device(A::ROCArray) -> ROCDevice

Return the device associated with the array A.

AMDGPU.device! — Function

device!(device::ROCDevice)

Switch current device being used. This switches only for a task inside which it is called.

Note

To select default device that will be used when creating new tasks, refer to default_device! for that.

AMDGPU.default_device — Function

default_device()::ROCDevice

Default device which will be used by default in tasks. Meaning when a task is created, it selects this device as default.

All subsequent uses rely on device() for device selection.

AMDGPU.default_device! — Function

default_device!(device::ROCDevice)

Set default device that will be used when creating new tasks.

Note

This does not change current device being used. Refer to device! for that.

AMDGPU.default_device_id — Function

default_device_id(kind::Symbol=:gpu) -> Int

Returns the numeric ID of the current default device, which is in the range of 1:length(AMDGPU.devices(kind)). This number should be stable for all processes on the same node, so long as any device filtering is consistently applied (such as ROCR_VISIBLE_DEVICES). The default_device_id! function accepts the same numeric ID that is produced by this function.

AMDGPU.default_device_id! — Function

default_device_id!(idx::Integer, kind::Symbol=:gpu)

Sets the default device to AMDGPU.devices(kind)[idx]. See default_device_id for details on the numbering semantics.