Devices

In AMDGPU, all GPU devices are auto-detected by the runtime, if they're supported.

AMDGPU maintains a global default device. The default device is relevant for all kernel and GPUArray operations. If one is not specified via @roc or an equivalent interface, then the default device is used for those operations, which affects compilation and kernel launch.

Task-Local Storage

AMDGPU.jl relies on Task-Local Storage, this means that default devices are default only within a given task. Other tasks migh have different default devices if user switched them.

The device bound to a current Julia task is accessible via AMDGPU.device(). The list of available devices can be queried with AMDGPU.devices.

If you have a HIPDevice object, you can also switch the device with AMDGPU.device!. This will switch it only within the task it is called from.

xd1 = AMDGPU.ones(Float32, 16) # On `AMDGPU.device()` device.

AMDGPU.device!(AMDGPU.devices()[2]) # Switch to second device.
xd2 = AMDPGU.ones(Float32, 16) # On second device.

To select a default device for newly created tasks, use AMDGPU.default_device!.

AMDGPU.default_device!(AMDGPU.devices()[3]) # New tasks will use 3rd device by default.
Threads.@spawn begin
    x = AMDGPU.ones(Float32, 16) # On third device.
    return
end

Additionally, devices have an associated numeric ID. The default device ID can be queried with AMDGPU.default_device_id, which returns an Int. This value is bounded between 1 and length(AMDGPU.devices()), and device 1 is the default device when AMDGPU is first loaded. The ID of the device associated with the current task can be queried with AMDGPU.device_id and changed with AMDGPU.device_id!.

AMDGPU.devices — Function

devices()

Get list of all devices.

AMDGPU.device — Function

device()::HIPDevice

Get currently active device. This device is used when launching kernels via @roc.

device(A::ROCArray) -> HIPDevice

Return the device associated with the array A.

AMDGPU.device! — Function

device!(device::HIPDevice)

Switch current device being used. This switches only for a task inside which it is called.

Note

To select default device that will be used when creating new tasks, refer to default_device! for that.

AMDGPU.default_device — Function

default_device()::HIPDevice

Default device which will be used by default in tasks. Meaning when a task is created, it selects this device as default.

All subsequent uses rely on device() for device selection.

AMDGPU.default_device! — Function

default_device!(device::HIPDevice)

Set default device that will be used when creating new tasks.

Note

This does not change current device being used. Refer to device! for that.

AMDGPU.device_id — Function

device_id(device::HIPDevice) -> Int

Returns the numerical device ID for device. See default_device_id for details on the numbering semantics.

AMDGPU.device_id! — Function

device_id!(idx::Integer, kind::Symbol=:gpu)

Sets the current device to AMDGPU.devices(kind)[idx]. See device_id for details on the numbering semantics.

AMDGPU.default_device_id — Function

default_device_id() -> Int

Returns the numeric ID of the current default device, which is in the range of 1:length(AMDGPU.devices()). This number should be stable for all processes on the same node, The default_device_id! function accepts the same numeric ID that is produced by this function.

AMDGPU.default_device_id! — Function

default_device_id!(idx::Integer, kind::Symbol=:gpu)

Sets the default device to AMDGPU.devices(kind)[idx]. See default_device_id for details on the numbering semantics.